chainhash: bump to v2
This bumps the major module version for chaincfg/chainhash to v2.  This
is needed to resolve an ambiguous import issue with the same package
existing in older revisions of the main btcd module.  Users who import
both a tagged revision of btcd prior to the chainhash module, as well as
another module such as btcd that requires the chainhash module may
be met with an ambiguous import error from the go tooling.